refactor code;

+++ backend/src/main/java/com/com/backend/controller/UserController.java
@@ -1,8 +1,11 @@
 package com.com.backend.controller;
 
 import com.com.backend.dto.Mail;
-import com.com.backend.dto.UsersDto;
+import com.com.backend.dto.request.UserUpdateRequest;
+import com.com.backend.dto.response.UserResponse;
+import com.com.backend.dto.response.UserResponseWithAbstracts;
 import com.com.backend.exception.AppException;
+import com.com.backend.model.enums.ExceptionType;
 import com.com.backend.service.EmailService;
 import com.com.backend.service.UsersService;
 import org.springframework.http.ResponseEntity;
@@ -29,39 +32,56 @@ public class UserController {
     @GetMapping
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<?> getAll(){
-        List<UsersDto> getAll = usersService.getAll();
+        List<UserResponse> getAll = usersService.getAll();
         return ResponseEntity.ok(getAll);
     }
 
     @GetMapping("/{id}")
-    @PreAuthorize("hasAnyRole('ACTIVE_PARTICIPANT', 'PASSIVE_PARTICIPANT')")
-    public ResponseEntity<UsersDto> getUser(@Valid @PathVariable Long id) throws AppException {
-        UsersDto users = usersService.getOne(id);
+    @PreAuthorize("hasAnyRole('ACTIVE_PARTICIPANT', 'PASSIVE_PARTICIPANT', 'ADMIN')")
+    public ResponseEntity<UserResponse> getUser(@Valid @PathVariable Long id) throws AppException {
+        UserResponse users = usersService.getOne(id);
         return ResponseEntity.ok(users);
     }
 
     @GetMapping("/{id}/admin")
-    @PreAuthorize("hasAnyRole('ADMIN')")
-    public ResponseEntity<UsersDto> getUserForAdmin(@Valid @PathVariable Long id) throws AppException {
-        UsersDto users = usersService.getOneForAdmin(id);
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<UserResponseWithAbstracts> getUserForAdmin(@Valid @PathVariable Long id) throws AppException {
+        UserResponseWithAbstracts users = usersService.getOneForAdmin(id);
         return ResponseEntity.ok(users);
     }
 
     @PutMapping("/{id}")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_ACTIVE_PARTICIPANT', 'ROLE_PASSIVE_PARTICIPANT')")
-    public ResponseEntity<?> updateUser(@Valid @PathVariable Long id,@PathParam("isUpdated") String isUpdated,
-                                        @Valid @RequestBody UsersDto usersDto) throws AppException {
+    public ResponseEntity<?> updateUser(@RequestHeader(value = "Authorization") String token,
+                                        @Valid @PathVariable Long id,@PathParam("isUpdated") String isUpdated,
+                                        @Valid @RequestBody UserUpdateRequest usersDto) throws AppException {
+
+        String email = usersService.getEmailFromToken(token);
+
+        /*if(!usersService.existsUserByEmail(usersDto.getEmail())) {
+            throw new AppException(EntityType.USER, ExceptionType.NOT_FOUND);
+        }*/
+
+        if(!usersService.getEmailFromUserId(usersDto.getId()).equals(email)) {
+            throw new AppException(ExceptionType.PRIVILEGES);
+        }
 
         if(isUpdated.equals("false"))
             return ResponseEntity.ok(usersDto);
 
-        UsersDto user = usersService.updateUser(usersDto);
+        UserResponse user = usersService.updateUser(usersDto);
         return ResponseEntity.ok(user);
     }
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_ACTIVE_PARTICIPANT', 'ROLE_PASSIVE_PARTICIPANT')")
-    public ResponseEntity<Void> deleteUser(@Valid @PathVariable Long id) throws AppException {
+    public ResponseEntity<Void> deleteUser(@RequestHeader(value = "Authorization") String token,
+                                           @Valid @PathVariable Long id) throws AppException {
+        String email = usersService.getEmailFromToken(token);
+        if(!usersService.getUserIdByEmail(email).equals(id)) {
+            throw new AppException(ExceptionType.PRIVILEGES);
+        }
+
         usersService.deleteUser(id);
         return ResponseEntity.ok().build();
     }
